Refactor UI and telemetry handling
- UPDATED TO LVGL 9.3.0
- Updated backlight constants to use a float value for BACKLIGHT_ACTIVE.
- Modified the messageReceived function to accept esp_now_recv_info for better context handling.
- Introduced telemetryDirty flag in Receiver struct to manage UI updates based on incoming telemetry.
- Enhanced the loop function to apply telemetry-driven UI updates efficiently.
- Refactored screen creation functions to use lv_button_create and lv_image_create for button and image objects.
- Added a new style for tab views and applied it to the main screen.
- Changed operating hours display functions to use unsigned long for better compatibility.
- Cleaned up style management functions for better readability and maintainability.

+++ src/uitools.c
@@ -68,19 +68,19 @@ float get_var_brightness_led3() { return brightness_led3; }
 
 void set_var_operating_hours1(uint32_t value) {
     if (objects.label_operating_hours_1 != NULL) {
-        lv_label_set_text_fmt(objects.label_operating_hours_1, "%u h", value);
+        lv_label_set_text_fmt(objects.label_operating_hours_1, "%lu h", (unsigned long)value);
     }
 }
 
 void set_var_operating_hours2(uint32_t value) {
     if (objects.label_operating_hours_2 != NULL) {
-        lv_label_set_text_fmt(objects.label_operating_hours_2, "%u h", value);
+        lv_label_set_text_fmt(objects.label_operating_hours_2, "%lu h", (unsigned long)value);
     }
 }
 
 void set_var_operating_hours3(uint32_t value) {
     if (objects.label_operating_hours_3 != NULL) {
-        lv_label_set_text_fmt(objects.label_operating_hours_3, "%u h", value);
+        lv_label_set_text_fmt(objects.label_operating_hours_3, "%lu h", (unsigned long)value);
     }
 }
